requirements.txt
playwright==1.55.0
pytest==8.4.2
pytest-playwright==0.7.1
allure-pytest==2.15.0
allure-python-commons~=2.15.0